backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.7.1
google-cloud-storage==3.8.0
google-generativeai>=0.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.11.7
pydantic-settings==2.12.0
pillow==12.1.0
